% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mouse_ID.R
\name{add_mouse_ID}
\alias{add_mouse_ID}
\title{add_mouse_ID}
\usage{
add_mouse_ID(x)
}
\arguments{
\item{x}{dataframe to add column to.}
}
\value{
The same data frame with mouseID column.
}
\description{
Adds mouseID to the original dataframe provided
}
\examples{
library(shiny)
library(shinythemes)
library(mousebreedeR)
library(dqrng)
library(dplyr)
library(gtools)
library(ggplot2)
library(reshape2)
library(viridis)
library(ggpubr)

head(exampleexampleData) # let's take a look at some example data before we run add_mouse_ID()
add_mouse_ID(exampleexampleData) # check it out! We have a new column!
}
