% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addmouseID.R
\name{addmouseID}
\alias{addmouseID}
\title{addmouseID}
\usage{
addmouseID(x)
}
\arguments{
\item{x}{dataframe to add column to.}
}
\value{
The same data frame with mouseID column.
}
\description{
Adds mouseID to the original dataframe provided
}
\examples{
addmouseID(exampleData.)
}
