% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpoints.R
\name{addpoints}
\alias{addpoints}
\title{addpoints}
\usage{
addpoints(x, desiredvector = desiredvec)
}
\arguments{
\item{x}{dataframe to assess points}

\item{desiredvector}{character vector of desired genotype.}
}
\value{
the number of points
}
\description{
Part of the internal functions to determine how much points a pup is worth.
}
\examples{
addpoints(x,desiredvector=exampledesiredvec)
}
