% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canwegetalltheallelesfromonecross.R
\name{canwegetalltheallelesfromonecross}
\alias{canwegetalltheallelesfromonecross}
\title{canwegetalltheallelesfromonecross}
\usage{
canwegetalltheallelesfromonecross(x, desiredvector = desiredvec)
}
\arguments{
\item{x}{the output of fertilize()}

\item{desiredvector}{a character vector of the desired genotype}
}
\value{
If a dataframe was supplied, it should also return a dataframe as well as a bar graph of the enriched pathways.
}
\description{
This function will tell you if multiple crosses will be required to get the desired mouse.
}
\examples{
canwegetalltheallelesfromonecross(x=examplefertilizeoutput,desiredvector=exampledesiredvec)
}
