% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_gametes.R
\name{compile_gametes}
\alias{compile_gametes}
\title{compile_gametes}
\usage{
compile_gametes(x)
}
\arguments{
\item{x}{the output of engage_in_meiosis()}
}
\value{
a dataframe of gametes
}
\description{
This finishes the create of the gametes.
}
\examples{
library(shiny)
library(shinythemes)
library(mousebreedeR)
library(dqrng)
library(dplyr)
library(gtools)
library(ggplot2)
library(reshape2)
library(viridis)
library(ggpubr)

head(exampleexampleData) # let's take a look at our example data
engage_in_meiosis_output_example<-engage_in_meiosis(exampleexampleData) # Check it out! We can look the gene makeup of eggs and sperm now
compile_gametes(engage_in_meiosis_output_example) # Meiosis step completed. Here are all the possible gametes from our breeder mice.
}
