% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engage_in_meiosis.R
\name{engage_in_meiosis}
\alias{engage_in_meiosis}
\title{engage_in_meiosis}
\usage{
engage_in_meiosis(x)
}
\arguments{
\item{x}{data frame with first columns as genotypes, and last rightmost column as sex column. Genotype columns should be numeric, and sex column should be M for male or F for female.}
}
\value{
First step of compiling the gametes, which is completed by the next function compilegamets.
}
\description{
The creation of gametes is one of the most beautiful adaptations in all biology. Watch it unfold.
}
\examples{
library(shiny)
library(shinythemes)
library(mousebreedeR)
library(dqrng)
library(dplyr)
library(gtools)
library(ggplot2)
library(reshape2)
library(viridis)
library(ggpubr)

head(exampleexampleData) # let's take a look at our example data
engage_in_meiosis(exampleexampleData) # Check it out! We have begun the process of engaging in meiosis.
}
