% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engageinmeiosis.R
\name{engageinmeiosis}
\alias{engageinmeiosis}
\title{engageinmeiosis}
\usage{
engageinmeiosis(x)
}
\arguments{
\item{x}{data frame with first columns as genotypes, and last rightmost column as sex column.}
}
\value{
First step of compiling the gametes, which is completed by the next function compilegamets.
}
\description{
Pathway analysis of DESeq2 result or character vector of differentially expressed genes which also plots results.
}
\examples{
engageinmeiosis(exampleData)
}
