% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_per_pup.R
\name{points_per_pup}
\alias{points_per_pup}
\title{points_per_pup}
\usage{
points_per_pup(x, desiredvector = desiredvec)
}
\arguments{
\item{x}{summarizepotentialpupoutput}

\item{desiredvector}{a vector of desired genotype.}
}
\value{
returns the dataframe which now has points per pups.
}
\description{
Calculates how many points a potential pup would be worth according to its genotype.
}
\examples{
points_per_pup(x=examplesummarizepotentialpupoutput,desiredvector=exampledesiredvec)
}
