% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointsperpup.R
\name{pointsperpup}
\alias{pointsperpup}
\title{pointsperpup}
\usage{
pointsperpup(x, desiredvector = desiredvec)
}
\arguments{
\item{x}{summarizepotentialpupoutput}

\item{desiredvector}{a vector of desired genotype.}
}
\value{
returns the dataframe which now has points per pups.
}
\description{
Calculates how many points a potential pup would be worth according to its genotype.
}
\examples{
pointsperpup(x=examplesummarizepotentialpupoutput,desiredvector=exampledesiredvec)
}
