% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_points_from_desired_outcome.R
\name{possible_points_from_desired_outcome}
\alias{possible_points_from_desired_outcome}
\title{possible_points_from_desired_outcome}
\usage{
possible_points_from_desired_outcome(desiredvector = desiredvec)
}
\arguments{
\item{desiredvector}{character vector of desired genotype.}
}
\value{
The number of possible points your desired pup is worth.
}
\description{
Calculates how many points any pup could be worth based on how many points your desired pup could be worth.
}
\examples{
possible_points_from_desired_outcome(desiredvector = exampledesiredvec)
}
