% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possiblepointsfromdesiredoutcome.R
\name{possiblepointsfromdesiredoutcome}
\alias{possiblepointsfromdesiredoutcome}
\title{possiblepointsfromdesiredoutcome}
\usage{
possiblepointsfromdesiredoutcome(desiredvector = desiredvec)
}
\arguments{
\item{desiredvector}{character vector of desired genotype.}
}
\value{
The number of possible points your desired pup is worth.
}
\description{
Calculates how many points any pup could be worth based on how many points your desired pup could be worth.
}
\examples{
possiblepointsfromdesiredoutcome(desiredvector = exampledesiredvec)
}
