% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sperm_and_eggs.R
\name{sperm_and_eggs}
\alias{sperm_and_eggs}
\title{sperm_and_eggs}
\usage{
sperm_and_eggs(x, sex = "sex")
}
\arguments{
\item{x}{the outpout of compile_gametes()}

\item{sex}{the column name in your compilegametesoutput that refers to sex.}
}
\value{
The possible sperm and eggs you could produce.
}
\description{
Pulls out sperm and eggs from the gametes created by compilegametes.
}
\examples{
sperm_and_eggs(x=examplecompilegametesoutput,sex='sex')
}
