% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizefertilization.R
\name{summarizefertilization}
\alias{summarizefertilization}
\title{summarizefertilization}
\usage{
summarizefertilization(x)
}
\arguments{
\item{x}{the output of fertilize()}
}
\value{
Dataframe with columns to summarize the probabalistic outcome of fertilizations, organized per cross and per locus.
}
\description{
summarizes the overall picture of fertilization.
}
\examples{
summarizefertilization(fertilizeoutput)
}
