% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizepotentialpups.R
\name{summarizepotentialpups}
\alias{summarizepotentialpups}
\title{summarizepotentialpups}
\usage{
summarizepotentialpups(x)
}
\arguments{
\item{x}{the output of fertilize()}
}
\value{
If a dataframe was supplied, it should also return a dataframe as well as a bar graph of the enriched pathways.
}
\description{
Will describe frequency of all potential pups from all potential crosses.
}
\examples{
summarizepotentialpups(fertilizeoutput)
}
