% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_pairs_should_i_breed.R
\name{which_pairs_should_i_breed}
\alias{which_pairs_should_i_breed}
\title{which_pairs_should_i_breed}
\usage{
which_pairs_should_i_breed(x = pointsperpupoutput, desiredvector = desiredvec)
}
\arguments{
\item{x}{the output of the points_per_pup function}

\item{desiredvector}{a character vector of the genotype you want to get to}
}
\value{
the crosses that produce the highest probability of desired pup or pups that are closest to the desired genotype.
}
\description{
This function assists the user in deciding which breeding scheme will produce the result most likely to give them the mouse of their preferred genotype
}
\examples{
which_pairs_should_i_breed(x=examplepointsperpupoutput,desiredvector=exampledesiredvec)
}
