% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichpairsshouldibreed.R
\name{whichpairsshouldibreed}
\alias{whichpairsshouldibreed}
\title{whichpairsshouldibreed}
\usage{
whichpairsshouldibreed(x = pointsperpupoutput, desiredvector = desiredvec)
}
\arguments{
\item{x}{the output of the pointsperpup function}

\item{desiredvector}{a character vector of the genotype you want to get to}
}
\value{
If a dataframe was supplied, it should also return a dataframe as well as a bar graph of the enriched pathways.
}
\description{
This function assists the user in deciding which breeding scheme will produce the result most likely to give them the mouse of their preferred genotype
}
\examples{
whichpairsshouldibreed(x=pointsperpupoutput,desiredvector=desiredvec)
}
