% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_points.R
\name{add_points}
\alias{add_points}
\title{add_points}
\usage{
add_points(x, desiredvector = desiredvec)
}
\arguments{
\item{x}{dataframe to assess points}

\item{desiredvector}{character vector of desired genotype.}
}
\value{
the number of points
}
\description{
Part of the internal functions to determine how much points a pup is worth.
}
\examples{
add_points(x,desiredvector=exampledesiredvec)
}
