% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_gametes.R
\name{compile_gametes}
\alias{compile_gametes}
\title{compile_gametes}
\usage{
compile_gametes(x)
}
\arguments{
\item{x}{the output of engage_in_meiosis()}
}
\value{
a dataframe of gametes
}
\description{
This finishes the create of the gametes.
}
\examples{
compile_gametes(examplemeiosisoutput)
}
