% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engageinmeiosis.R
\name{engageinmeiosis}
\alias{engageinmeiosis}
\title{engageinmeiosis}
\usage{
engageinmeiosis(x)
}
\arguments{
\item{x}{data frame with first columns as genotypes, and last rightmost column as sex column. Genotype columns should be numeric, and sex column should be M for male or F for female.}
}
\value{
First step of compiling the gametes, which is completed by the next function compilegamets.
}
\description{
The creation of gametes is one of the most beautiful adaptations in all biology. Watch it unfold.
}
\examples{
engageinmeiosis(exampleData)
}
