% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertilize.R
\name{fertilize}
\alias{fertilize}
\title{fertilize}
\usage{
fertilize(malegametes = sperm, femalegametes = eggs)
}
\arguments{
\item{malegametes}{the sperm from spermandeggs() function}

\item{femalegametes}{the eggs from spermandeggs() function}
}
\value{
If a dataframe was supplied, it should also return a dataframe as well as a bar graph of the enriched pathways.
}
\description{
Pathway analysis of DESeq2 result or character vector of differentially expressed genes which also plots results.
}
\examples{
fertilize(malegametes = sperm,femalegametes = eggs)
}
