% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertilize.R
\name{fertilize}
\alias{fertilize}
\title{fertilize}
\usage{
fertilize(malegametes = sperm, femalegametes = eggs)
}
\arguments{
\item{malegametes}{the sperm from sperm_and_eggs() function}

\item{femalegametes}{the eggs from sperm_and_eggs() function}
}
\value{
Dataframe of possible pups.
}
\description{
Watch as sperm and egg connect and create a life form from the combination of their genetic material.
}
\examples{
fertilize(malegametes = examplesperm,femalegametes = exampleeggs)
}
