% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sperm_and_eggs.R
\name{sperm_and_eggs}
\alias{sperm_and_eggs}
\title{sperm_and_eggs}
\usage{
sperm_and_eggs(x, sex = "sex")
}
\arguments{
\item{x}{the outpout of compile_gametes()}

\item{sex}{the column name in your compilegametesoutput that refers to sex.}
}
\value{
The possible sperm and eggs you could produce.
}
\description{
Pulls out sperm and eggs from the gametes created by compilegametes.
}
\examples{
library(shiny)
library(shinythemes)
library(mousebreedeR)
library(dqrng)
library(dplyr)
library(gtools)
library(ggplot2)
library(reshape2)
library(viridis)
library(ggpubr)
head(exampleexampleData) # let's take a look at our example data
engage_in_meiosis_output_example<-engage_in_meiosis(exampleexampleData) # Check it out! We can look the gene makeup of eggs and sperm now
compile_gametes_output_example<-compile_gametes(engage_in_meiosis_output_example) # Meiosis step completed. Here are all the possible gametes from our breeder mice.
sperm_and_eggs(x=compile_gametes_output_example,sex='sex') # Saves the outputs of which gametes are sperm, and which are eggs.
}
