% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_fertilization.R
\name{summarize_fertilization}
\alias{summarize_fertilization}
\title{summarize_fertilization}
\usage{
summarize_fertilization(x)
}
\arguments{
\item{x}{the output of fertilize()}
}
\value{
Dataframe with columns to summarize the probabalistic outcome of fertilizations, organized per cross and per locus.
}
\description{
summarizes the overall picture of fertilization.
}
\examples{
summarize_fertilization(examplefertilizeoutput)
}
