% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_potential_pups.R
\name{summarize_potential_pups}
\alias{summarize_potential_pups}
\title{summarize_potential_pups}
\usage{
summarize_potential_pups(x)
}
\arguments{
\item{x}{the output of fertilize()}
}
\value{
Dataframe with probability information of each corss, organized to per pup created.
}
\description{
Will describe frequency of all potential pups from all potential crosses.
}
\examples{
summarize_potential_pups(examplefertilizeoutput)
}
