% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizefertilization.R
\name{summarizefertilization}
\alias{summarizefertilization}
\title{summarizefertilization}
\usage{
summarizefertilization(x)
}
\arguments{
\item{x}{the output of fertilize()}
}
\value{
If a dataframe was supplied, it should also return a dataframe as well as a bar graph of the enriched pathways.
}
\description{
summarizes the overall picture of fertilization.
}
\examples{
summarizefertilization(fertilizeoutput)
}
