% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizepotentialpups.R
\name{summarizepotentialpups}
\alias{summarizepotentialpups}
\title{summarizepotentialpups}
\usage{
summarizepotentialpups(x)
}
\arguments{
\item{x}{the output of fertilize()}
}
\value{
Dataframe with probability information of each corss, organized to per pup created.
}
\description{
Will describe frequency of all potential pups from all potential crosses.
}
\examples{
summarizepotentialpups(fertilizeoutput)
}
