% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_pairs_should_i_breed.R
\name{which_pairs_should_i_breed}
\alias{which_pairs_should_i_breed}
\title{which_pairs_should_i_breed}
\usage{
which_pairs_should_i_breed(x = pointsperpupoutput, desiredvector = desiredvec)
}
\arguments{
\item{x}{the output of the points_per_pup function}

\item{desiredvector}{a character vector of the genotype you want to get to}
}
\value{
the crosses that produce the highest probability of desired pup or pups that are closest to the desired genotype.
}
\description{
This function assists the user in deciding which breeding scheme will produce the result most likely to give them the mouse of their preferred genotype
}
\examples{
library(shiny)
library(shinythemes)
library(mousebreedeR)
library(dqrng)
library(dplyr)
library(gtools)
library(ggplot2)
library(reshape2)
library(viridis)
library(ggpubr)
head(exampleexampleData) # let's take a look at our example data
# Check it out! We can look the gene makeup of eggs and sperm now
engage_in_meiosis_output_example<-engage_in_meiosis(exampleexampleData) 
# Meiosis step completed. Here are all the possible gametes from our breeder mice.
compile_gametes_output_example<-compile_gametes(engage_in_meiosis_output_example) 
# Saves the outputs of which gametes are sperm, and which are eggs.
sperm_and_eggs_output_example<-sperm_and_eggs(x=compile_gametes_output_example,sex='sex') 
# create all potential pups from all possible pairings.
fertilize_output_example<-fertilize(malegametes = sperm,femalegametes = eggs) 
fertilize_output<-fertilize_output_example
desiredvec<-c('het','het','het','het') # the genotype of your desired mouse
# take a look at the distributions of potential pups
summarize_potential_pups_output<-summarize_potential_pups(fertilize_output_example) 
points_per_pup_output_example<-points_per_pup(x=summarize_potential_pups_output,desiredvector=desiredvec)
which_pairs_should_i_breed(x=points_per_pup_output_example,desiredvector=desiredvec)
}
